+++ include/internal/std/Enums.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <internal/SerpentLua.hpp>
+#include <sol/sol.hpp>
+
+namespace SerpentLua::internal::ScriptBuiltin::Enums {
+	sol::table entry(sol::state_view ts);
+
+	struct UI {
+		enum class NodeType {
+			Node, // CCNode
+			Sprite, // CCSprite
+			Button, // CCMenuItemExt
+			Label, // CCLabelBMFont
+			Menu // CCMenu
+		};
+	};
+};